岗位自检： 1.按批次生成，2.测试

+++ src/main/java/me/jinkun/rds/sys/dao/GwQualityInspectTaskDetailMapper.java
@@ -14,7 +14,7 @@ public interface GwQualityInspectTaskDetailMapper {
 
     int insert(GwQualityInspectTaskDetail record);
 
-    int insertSelective(GwQualityInspectTaskDetail record);
+    long insertSelective(GwQualityInspectTaskDetail record);
 
     List<GwQualityInspectTaskDetail> selectByExample(GwQualityInspectTaskDetailExample example);
